feat: split forum topic subscription tasks
